pyproject.toml
[tool.poetry]
name = "neural_canvas"
version = "0.1.0"
description = "creative-deep-learning-through-implicit-neural-representations"
authors = ["neale <nealeratzlaff@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
torch = "^2.0.0"
torchvision = "^0.15.1"
networkx = "^3.1"
opencv-python = "^4.7.0.72"
einops = "^0.6.0"
tifffile = "^2023.4.12"
moviepy = "^1.0.3"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"
pyyaml = "^6.0"
matplotlib = "^3.7.1"

[tool.poetry.group.dev.dependencies]
clip = "^0.2.0"
openai = "^0.27.4"
requests = "^2.28.2"
tqdm = "^4.65.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
